+++ src/main/java/bridge/domain/BridgeGame.java
@@ -1,10 +1,12 @@
-package bridge;
+package bridge.domain;
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
  */
 public class BridgeGame {
 
+    public void generateBridge() {
+    }
     /**
      * 사용자가 칸을 이동할 때 사용하는 메서드
      * <p>
