+++ src/main/java/bridge/domain/BridgeMaker.java
@@ -1,4 +1,6 @@
-package bridge;
+package bridge.domain;
+
+import bridge.BridgeNumberGenerator;
 
 import java.util.List;
